\docType{data}
\name{Big5}
\alias{Big5}
\title{big 5 personality traits}
\format{a \code{data.frame} instance, 1 row per item, 10 or 11 items per scale.}
\source{
\url{http://en.wikipedia.org/wiki/Big_Five_personality_traits}
}
\usage{
Big5
}
\description{
This data set contains items testing, the big 5 personality
traits.
}
\examples{
data(Big5)
str(Big5)
}

