\docType{data}
\name{eAnswers}
\alias{eAnswers}
\title{norm: items aggregated within scales}
\format{a \code{data.frame} instance, 1 row per item (115 total).}
\source{
  simulated data set
}
\description{
  This data set contains an example responses to a launched
  questionnaire object. The data would be the output from
  the function calls questionnaire(), launch() and the
  responses assumed to be stored in an object with
  get.answers().
}
\examples{
data( eAnswers)
str(eAnswers)
}
\keyword{datasets}

