\docType{data}
\name{eKey}
\alias{eKey}
\title{key: item-scale mapping}
\format{a \code{data.frame} instance, 1 row per item, 5 items per scale, 23 scales.}
\source{
simulated data set
}
\usage{
eKey
}
\description{
This data set contains the mapping from questions (items)
to scale (constructs)
}
\examples{
data(eKey)
str(eKey)
}

