\docType{data}
\name{ExampleKey}
\alias{ExampleKey}
\title{key: item-scale mapping}
\format{a \code{data.frame} instance, 1 row per item, 5 items per scale, 23 scales.}
\source{
  simulated data set
}
\description{
  This data set contains the mapping from questions (items)
  to scale (constructs)
}
\keyword{datasets}

