\docType{data}
\name{ExampleNorm}
\alias{ExampleNorm}
\title{norm: items aggregated within scales}
\format{a \code{data.frame} instance, 1 row per scale (23 total), T-score and mean.rating}
\source{
  simulated data set
}
\description{
  This data set contains the mean rating and T-score within
  each scale
}
\keyword{datasets}

