\name{get.answers}
\alias{get.answers}
\title{extracts responses from a questionnaire object thats has been lauched and answered}
\usage{
  get.answers(q.frame = q.frame)
}
\arguments{
  \item{q.frame}{rJava object that has been given back from
  launch() function}

  \item{verbose}{Print information during execution}
}
\value{
  answers a data frame of recorded answers for each item
}
\description{
  extracts responses from a questionnaire object thats has
  been lauched and answered
}

