\name{launch}
\alias{launch}
\title{launches a survey, that is displays the itmes and the answer scale and records the responses
from a questionnaire object}
\usage{
  launch(subject.id = "id01", session.id = "01",
    subject.group = "PILOT", qstnr = "test")
}
\arguments{
  \item{qstnr}{object of class questionnaire}
}
\value{
  S3 \code{rJava} object; a list consisting of ...
}
\description{
  launches a survey, that is displays the itmes and the
  answer scale and records the responses from a
  questionnaire object
}

