\name{make.scale}
\alias{make.scale}
\title{makes an xml nodes for the answer-scales}
\usage{
make.scale(scale, labels, reversed = F, scalenumber = 1, questionnaire)
}
\arguments{
  \item{scale}{}

  \item{labels}{}

  \item{reversed=F}{}

  \item{scalenumber=1}{}

  \item{questionnaire}{}
}
\value{
S3 \code{XML} object
}
\description{
makes an xml nodes for the answer-scales
}

