% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQuestionnaireTag.R
\name{makeQuestionnaireTag}
\alias{makeQuestionnaireTag}
\title{Build a questionnaire tag}
\usage{
makeQuestionnaireTag()
}
\description{
Build the questionnaire tag, ready to be inserted in a HTML document.
}
