\name{print.questionnaire}
\alias{print.questionnaire}
\title{prints a questionnaire object into the console}
\usage{
\method{print}{questionnaire}(x, ...)
}
\arguments{
  \item{x}{object of class questionnaire}
}
\value{
console formated text text
}
\description{
prints a questionnaire object into the console
}

